Allow EmailAddress model override
