+++ simple_email_confirmation/admin.py
@@ -1,10 +1,11 @@
 from django.contrib import admin
 
-from .models import EmailAddress
+from .settings import EmailAddressModel
 
 
 class EmailAddressAdmin(admin.ModelAdmin):
     list_display = ('user', 'email', 'key', 'set_at', 'confirmed_at')
     search_fields = ('email', 'key')
 
-admin.site.register((EmailAddress,), EmailAddressAdmin)
+
+admin.site.register(EmailAddressModel, EmailAddressAdmin)
